ECL-736: Added Amended Answers section on view and pdf when getEclReturnEnabled is ON (#128)
* ECL-736: Work in progress

* ECL-736: WIP

* ECL-736: WIP

Functionality works, although some tests have been commented out to test functionality at runtime. I'm also in the middle of finishing the trait unit tests. UK revenue val is currently broken.

* ECL-736: Moved a function from viewmodels into trait

* ECL-736: Added skeleton of remaining tests and begun fixing logic of track changes boolean methods

* ECL-736: Added remaining tests for track changes relevantAp12Months and

* ECL-736: Completed tests for hasRelevantApLengthChanged and corrected logic

* ECL-736: Improved descriptions on tests, corrected logic on hasUkRevenueChanged, and added hasUkRevenueChanged tests

* ECL-736: Added content where there are no changes and tweaked aggregate boolean functions in track changes

* ECL-736: Finished tests and corrected logic for hasCarriedOutAmlRegulatedActivityForFullFyChanged

* ECL-736: Completed tests for hasAmlRegulatedActivityLengthChanged

* ECL-736: Corrected logic and finished tests for hasCalculatedBandSummaryChanged

* ECL-736: Finished remaining tests and logic changes for TrackEclReturnChangesSpec

* ECL-736: Added tests for additional values in TrackEclReturnChanges trait

* ECL-736: Corrected issue with switching feature flag on/off

* ECL-736: Corrected AmountDueControllerSpec tests

* ECL-736: Corrected tests in CheckYourAnswersController and added new tests to cover error scenarios

* ECL-736: Added and enhanced further tests

* ECL-736: Logic fix and added tests for FirstTimeReturn

* ECL-736: Hide amended answers on PDF when empty

* ECL-736: Brought pdfViewHtmlOrError and viewHtmlOrError into consistency

* ECL-736: Minor cleanup in a controller.

* ECL-736: scalfmt

* ECL-736: PR feedback

* ECL-736: PR feedback

* ECL-736: PR feedback. Simplified for comprehensions

+++ it/uk/gov/hmrc/economiccrimelevyreturns/UkRevenueISpec.scala
@@ -39,7 +39,7 @@ class UkRevenueISpec extends ISpecBase with AuthorisedBehaviour {
 
       val eclReturn           = random[EclReturn]
         .copy(amlRegulatedActivityLength = Some(365), relevantAp12Months = Some(true), calculatedLiability = None)
-      val ukRevenue           = bigDecimalInRange(UkRevenueThreshold, MinMaxValues.RevenueMax.toDouble).sample.get
+      val ukRevenue           = bigDecimalInRange(UkRevenueThreshold.toDouble, MinMaxValues.RevenueMax.toDouble).sample.get
       val calculatedLiability = random[CalculatedLiability].copy(calculatedBand = Large)
 
       stubGetReturn(eclReturn)
@@ -69,7 +69,7 @@ class UkRevenueISpec extends ISpecBase with AuthorisedBehaviour {
 
       val eclReturn           = random[EclReturn]
         .copy(amlRegulatedActivityLength = Some(365), relevantAp12Months = Some(true), calculatedLiability = None)
-      val ukRevenue           = bigDecimalInRange(MinMaxValues.RevenueMin.toDouble, UkRevenueThreshold).sample.get
+      val ukRevenue           = bigDecimalInRange(MinMaxValues.RevenueMin.toDouble, UkRevenueThreshold.toDouble).sample.get
       val calculatedLiability =
         random[CalculatedLiability].copy(calculatedBand = Small)
